Ensure control page loads and displays correct room temperature when in fahrenheit
This change overcomes a limitation in how large a payload we can send through the http server.  This sends the header, control and footer content in chunks.  I also noticed a bug from a previous commit where we were overwriting the room temperature converted to fahrenheit with the temperature before conversion right after.